Address committee review comments
Check number of eliminated classes in per class figure (fixed the figure bug).
Figures to be added after text not before or within (I changed the position to force it to be after the text direct)
Reduce Abstract
Add the results in the introduction
How we choose 4,7 Layers and unites
Talk about out of memory issue
Remove any future words about the current work (I did my best but I not sure if there is any other issues or not).

.